+++ packages/dash-table/src/dash-table/utils/actions.js
@@ -47,26 +47,25 @@ export function getAffectedColumns(column, columns, headerRowIndex, mergeDuplica
     );
 }
 
-export function clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
-    const rejectedColumnIds = getAffectedColumns(column, columns, headerRowIndex, mergeDuplicateHeaders);
-
-    return {
-        data: R.map(R.omit(rejectedColumnIds), data)
-    };
+export function clearColumn(column, columns, visibleColumns, headerRowIndex, mergeDuplicateHeaders, _data) {
+    const {data} = deleteColumn(column, columns, visibleColumns, headerRowIndex, mergeDuplicateHeaders, _data);
+    return {data};
 }
 
-export function deleteColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
-    const {groupIndexFirst, groupIndexLast} = getGroupedColumnIndices(
-        column, columns, headerRowIndex, mergeDuplicateHeaders, columns.indexOf(column)
+export function deleteColumn(column, columns, visibleColumns, headerRowIndex, mergeDuplicateHeaders, data) {
+    const rejectedColumnIds = getAffectedColumns(
+        column,
+        visibleColumns,
+        headerRowIndex,
+        mergeDuplicateHeaders
     );
 
     return {
-        columns: R.remove(
-            groupIndexFirst,
-            1 + groupIndexLast - groupIndexFirst,
+        columns: R.filter(
+            col => (rejectedColumnIds.indexOf(col.id) === -1),
             columns
         ),
-        ...clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data),
+        data: R.map(R.omit(rejectedColumnIds), data),
         // NOTE - We're just clearing these so that there aren't any
         // inconsistencies. In an ideal world, we would probably only
         // update them if they contained one of the columns that we're
@@ -106,7 +105,7 @@ export function changeColumnHeader(column, columns, headerRowIndex, mergeDuplica
     }
 
     const { groupIndexFirst, groupIndexLast } = getGroupedColumnIndices(
-        column, newColumns, headerRowIndex, mergeDuplicateHeaders, columnIndex
+        column, newColumns, headerRowIndex, mergeDuplicateHeaders, columnIndex, true
     );
 
     R.range(groupIndexFirst, groupIndexLast + 1).map(i => {
@@ -122,5 +121,8 @@ export function changeColumnHeader(column, columns, headerRowIndex, mergeDuplica
 
 export function editColumnName(column, columns, headerRowIndex, mergeDuplicateHeaders) {
     const newColumnName = window.prompt('Enter a new column name');
+    if (newColumnName === null) {
+        return null;
+    }
     return changeColumnHeader(column, columns, headerRowIndex, mergeDuplicateHeaders, newColumnName);
 }