+++ packages/dash-table/tests/cypress/tests/standalone/column_test.ts
@@ -27,6 +27,17 @@ describe(`column, mode=${AppMode.Actionable}, flavor=${AppFlavor.Merged}, ${AppF
         DashTable.getHeader(1, 1).within(() => cy.get('span.column-header-name').should('have.html', 'Paris'));
     });
 
+    it('keeps hidden pieces when deleting a merged column', () => {
+        cy.get('.column-4 .column-header--hide').click(); // Boston
+        cy.get('.column-2 .column-header--hide').click(); // Montréal
+        DashTable.deleteColumnById(0, 'ccc'); // City
+        cy.get('.show-hide').click();
+        cy.get('.show-hide-menu-item input').eq(1).click(); // Montréal
+        cy.get('.show-hide-menu-item input').eq(2).click(); // Boston
+        cy.get('.column-1 .column-header-name').eq(2).should('have.html', 'Montréal');
+        cy.get('.column-2 .column-header-name').eq(1).should('have.html', 'Boston');
+    });
+
     it('can clear column', () => {
         DashTable.getFilter(0).click();
         DOM.focused.type(`is num`);
@@ -60,6 +71,29 @@ describe(`column, mode=${AppMode.Actionable}, flavor=${AppFlavor.Merged}, ${AppF
         DashTable.getFilter(2).within(() => cy.get('input').should('have.value', ''));
         DashTable.getFilter(3).within(() => cy.get('input').should('have.value', 'is num'));
         DashTable.getFilter(4).within(() => cy.get('input').should('have.value', ''));
+    });
+
+    it('keeps hidden pieces when clearing a merged column', () => {
+        // initial state of city columns
+        DashTable.getCell(0, 1).within(() => cy.get('.dash-cell-value').should('have.html', '1'));
+        DashTable.getCell(0, 2).within(() => cy.get('.dash-cell-value').should('have.html', '100'));
+        DashTable.getCell(0, 3).within(() => cy.get('.dash-cell-value').should('have.html', '1'));
+        DashTable.getCell(0, 4).within(() => cy.get('.dash-cell-value').should('have.html', '2'));
+        DashTable.getCell(0, 5).within(() => cy.get('.dash-cell-value').should('have.html', '10'));
+
+        cy.get('.column-4 .column-header--hide').click(); // Boston
+        cy.get('.column-2 .column-header--hide').click(); // Montréal
+        DashTable.clearColumnById(0, 'ccc'); // City
+        cy.get('.show-hide').click();
+        cy.get('.show-hide-menu-item input').eq(2).click(); // Montréal
+        cy.get('.show-hide-menu-item input').eq(4).click(); // Boston
+
+        // after clear and re-show -> only Montréal and Boston have data
+        DashTable.getCell(0, 1).within(() => cy.get('.dash-cell-value').should('have.html', ''));
+        DashTable.getCell(0, 2).within(() => cy.get('.dash-cell-value').should('have.html', '100'));
+        DashTable.getCell(0, 3).within(() => cy.get('.dash-cell-value').should('have.html', ''));
+        DashTable.getCell(0, 4).within(() => cy.get('.dash-cell-value').should('have.html', '2'));
+        DashTable.getCell(0, 5).within(() => cy.get('.dash-cell-value').should('have.html', ''));
     });
 
     it('can hide column', () => {